refactor: 시험관련 엔티티 인덱스 최적화
