+++ src/main/java/com/airoom/airoom/board/entity/BoardType.java
@@ -1,4 +1,4 @@
-package com.airoom.airoom.board;
+package com.airoom.airoom.board.entity;
 
 public enum BoardType {
     GROUP,ASSIGN,SUBJECT,HOMEWORK
